+++ src/OmniSharp.Roslyn.CSharp/Services/Intellisense/IntellisenseService.cs
@@ -14,14 +14,14 @@ using OmniSharp.Roslyn.CSharp.Services.Documentation;
 
 namespace OmniSharp.Roslyn.CSharp.Services.Intellisense
 {
-    //[OmniSharpHandler(OmniSharpEndpoints.AutoComplete, LanguageNames.CSharp)]
-    public class IntellisenseOldService : IRequestHandler<AutoCompleteRequest, IEnumerable<AutoCompleteResponse>>
+    [OmniSharpHandler(OmniSharpEndpoints.AutoComplete, LanguageNames.CSharp)]
+    public class IntellisenseService : IRequestHandler<AutoCompleteRequest, IEnumerable<AutoCompleteResponse>>
     {
         private readonly OmniSharpWorkspace _workspace;
         private readonly FormattingOptions _formattingOptions;
 
         [ImportingConstructor]
-        public IntellisenseOldService(OmniSharpWorkspace workspace, FormattingOptions formattingOptions)
+        public IntellisenseService(OmniSharpWorkspace workspace, FormattingOptions formattingOptions)
         {
             _workspace = workspace;
             _formattingOptions = formattingOptions;
@@ -38,29 +38,72 @@ namespace OmniSharp.Roslyn.CSharp.Services.Intellisense
                 var sourceText = await document.GetTextAsync();
                 var position = sourceText.Lines.GetPosition(new LinePosition(request.Line, request.Column));
 
+                //var options = await document.GetOptionsAsync();
+                //options.WithChangedOption(new PerLanguageOption<SnippetsRule>("CompletionOptions", "SnippetsBehavior", SnippetsRule.AlwaysInclude), LanguageNames.CSharp, SnippetsRule.AlwaysInclude);
+                //options.WithChangedOption()
+                var semanticModel = await document.GetSemanticModelAsync();
                 var service = CompletionService.GetService(document);
                 var completionList = await service.GetCompletionsAsync(document, position);
+                //options: options.WithChangedOption(new PerLanguageOption<SnippetsRule>("CompletionOptions", "SnippetsBehavior", SnippetsRule.AlwaysInclude)));
 
                 // Add keywords from the completion list. We'll use the recommender service to get symbols
                 // to create snippets from.
+
+
+                var recommendedSymbols = await Recommender.GetRecommendedSymbolsAtPositionAsync(semanticModel, position, _workspace);
 
                 if (completionList != null)
                 {
                     foreach (var item in completionList.Items)
                     {
-                        if (item.Tags.Contains(CompletionTags.Keyword))
+                        //if (item.Tags.Contains(CompletionTags.Keyword))
                         {
                             // Note: For keywords, we'll just assume that the completion text is the same
                             // as the display text.
                             var keyword = item.DisplayText;
                             if (keyword.IsValidCompletionFor(wordToComplete))
                             {
+                                if (item.Properties.ContainsKey("Provider") && item.Properties["Provider"] == "Microsoft.CodeAnalysis.CSharp.Completion.Providers.SymbolCompletionProvider")
+                                {
+                                    //var symbolsTask = GetSymbolsAsync.Invoke(null, new object[] { item, document, default(CancellationToken) }) as Task<ImmutableArray<ISymbol>>;
+                                    //var symbols = await symbolsTask;
+                                    //SymbolFinder.
+                                    //foreach (var symbol in symbols.Where(s => s.Name.IsValidCompletionFor(wordToComplete)))
+                                    //service.GetDescriptionAsync
+                                    {
+                                        var symbols = recommendedSymbols.Where(x => x.Name == item.Properties["SymbolName"] && (int)x.Kind == int.Parse(item.Properties["SymbolKind"])).Distinct();
+
+                                        if (symbols != null && symbols.Any())
+                                        {
+                                            foreach (var symbol in symbols)
+                                            {
+                                                if (symbol != null)
+                                                {
+                                                    if (request.WantSnippet)
+                                                    {
+                                                        foreach (var completion in MakeSnippetedResponses(request, symbol))
+                                                        {
+                                                            completions.Add(completion);
+                                                        }
+                                                    }
+                                                    else
+                                                    {
+                                                        completions.Add(MakeAutoCompleteResponse(request, symbol));
+                                                    }
+                                                }
+                                            }
+
+                                            continue;
+                                        }
+                                    }
+                                }
+
                                 var response = new AutoCompleteResponse()
                                 {
                                     CompletionText = item.DisplayText,
                                     DisplayText = item.DisplayText,
                                     Snippet = item.DisplayText,
-                                    Kind = request.WantKind ? "Keyword" : null
+                                    Kind = request.WantKind ? item.Tags.First() : null
                                 };
 
                                 completions.Add(response);
@@ -69,23 +112,23 @@ namespace OmniSharp.Roslyn.CSharp.Services.Intellisense
                     }
                 }
 
-                var model = await document.GetSemanticModelAsync();
-                var symbols = await Recommender.GetRecommendedSymbolsAtPositionAsync(model, position, _workspace);
+                //var model = await document.GetSemanticModelAsync();
+                //var symbols = await Recommender.GetRecommendedSymbolsAtPositionAsync(model, position, _workspace);
 
-                foreach (var symbol in symbols.Where(s => s.Name.IsValidCompletionFor(wordToComplete)))
-                {
-                    if (request.WantSnippet)
-                    {
-                        foreach (var completion in MakeSnippetedResponses(request, symbol))
-                        {
-                            completions.Add(completion);
-                        }
-                    }
-                    else
-                    {
-                        completions.Add(MakeAutoCompleteResponse(request, symbol));
-                    }
-                }
+                //foreach (var symbol in symbols.Where(s => s.Name.IsValidCompletionFor(wordToComplete)))
+                //{
+                //    if (request.WantSnippet)
+                //    {
+                //        foreach (var completion in MakeSnippetedResponses(request, symbol))
+                //        {
+                //            completions.Add(completion);
+                //        }
+                //    }
+                //    else
+                //    {
+                //        completions.Add(MakeAutoCompleteResponse(request, symbol));
+                //    }
+                //}
             }
 
             return completions
